Added some sorting.

+++ src/com/jhalkjar/caoscomp/gui/Editor.java
@@ -1,9 +1,7 @@
 package com.jhalkjar.caoscomp.gui;
 
-import com.codename1.io.Log;
 import com.codename1.ui.*;
 import com.codename1.ui.layouts.BorderLayout;
-import com.codename1.ui.layouts.BoxLayout;
 import com.codename1.ui.plaf.Style;
 import com.codename1.ui.plaf.UIManager;
 import com.jhalkjar.caoscomp.backend.Rute;
@@ -21,9 +19,9 @@ public class Editor extends Form {
     Label l = new Label("Retrieving image..");
     Rute r;
     boolean edit;
-    EditorTool tool = new EditorTool() {
 
-    };
+    Slider sl = new Slider();
+    CheckBox delete = new CheckBox(FontImage.createMaterial(FontImage.MATERIAL_DELETE, s));
 
     Component editorBar;
 
@@ -32,7 +30,7 @@ public class Editor extends Form {
 
         r = rute;
         edit = r.getAuthor().equals(DB.getInstance().getLoggedInUser());
-
+        isLocal = r.isLocal();
 
         addOrientationListener(evt -> {
             removeComponent(editorBar);
@@ -52,14 +50,25 @@ public class Editor extends Form {
                 canvas = new Canvas(r.getPoints(), edit);
                 canvas.doSetImage(image);
                 canvas.addClickListener((x, y) -> {
-                    tool.OnClicked(x, y);
+                    Point p = new Point(x,y, (float) sl.getProgress() / 100.0f);
+                    r.getPoints().add(p);
+                    if(selected != null) selected.setSelected(false);
+                    selected = p;
+                    p.setSelected(true);
                     r.save();
                 });
                 canvas.addMoveListener(p -> {
-                    tool.OnMoved(p);
                     r.save();
                 });
-                canvas.addSelectionListener(p -> tool.OnSelected(p));
+                canvas.addSelectionListener(p -> {
+                    if(delete.isSelected()) r.getPoints().remove(p);
+                    else {
+                        sl.setProgress((int) (p.getSize()*100.0));
+                        if(selected != null) selected.setSelected(false);
+                        selected = p;
+                        p.setSelected(true);
+                    }
+                });
                 l.setHidden(true);
                 add(BorderLayout.CENTER, canvas);
                 revalidate();
@@ -77,8 +86,8 @@ public class Editor extends Form {
     }
 
     Component createEditorBar() {
-        CheckBox delete = new CheckBox(FontImage.createMaterial(FontImage.MATERIAL_DELETE, s));
-        Slider sl = new Slider();
+
+
         sl.setVertical(isVertical());
         sl.setMinValue(2);
         sl.setMaxValue(50);
@@ -94,29 +103,6 @@ public class Editor extends Form {
 
         });
 
-        tool = new EditorTool() {
-
-            public void OnClicked(float x, float y) {
-                Point p = new Point(x,y, (float) sl.getProgress() / 100.0f);
-                r.getPoints().add(p);
-                if(selected != null) selected.setSelected(false);
-                selected = p;
-                p.setSelected(true);
-
-            }
-
-            public void OnSelected(Point p) {
-                if(delete.isSelected()) r.getPoints().remove(p);
-                else {
-                    sl.setProgress((int) (p.getSize()*100.0));
-                    if(selected != null) selected.setSelected(false);
-                    selected = p;
-                    p.setSelected(true);
-
-                }
-            }
-
-        };
         Container c = new Container(new BorderLayout());
         c.add(BorderLayout.CENTER, sl);
         if(isVertical()) {
@@ -128,30 +114,33 @@ public class Editor extends Form {
         return c;
     }
 
-    void populateToolbar() {
+    boolean isLocal;
 
-        getToolbar().addCommandToLeftBar("", FontImage.createMaterial(FontImage.MATERIAL_ARROW_BACK, s), (e) -> {
+    void populateToolbar() {
+        Toolbar tb = new Toolbar();
+        setToolbar(tb);
+
+        tb.addCommandToLeftBar("", FontImage.createMaterial(FontImage.MATERIAL_ARROW_BACK, s), (e) -> {
             new RuteList().showBack();
         });
 
-//        MultiToggleButton<Boolean> multiToggleButton = new MultiToggleButton<>();
-//        multiToggleButton.addState(true, FontImage.createMaterial(FontImage.MATERIAL_STAR, s));
-//        multiToggleButton.addState(false, FontImage.createMaterial(FontImage.MATERIAL_STAR_BORDER, s));
-//        multiToggleButton.setCurrentState(r.isLocal());
-//        multiToggleButton.addActionListener(evt -> r.setLocal(multiToggleButton.getCurrentState()));
-//
-        Container cnt = new Container(BoxLayout.x());
-//        if(edit) {
-            Button b2 = new Button(FontImage.createMaterial(FontImage.MATERIAL_DELETE, s));
-            b2.addActionListener(evt -> {
+        char image = isLocal ? FontImage.MATERIAL_STAR : FontImage.MATERIAL_STAR_BORDER;
+        tb.addCommandToRightBar("", FontImage.createMaterial(image, s), evt -> {
+
+            r.setLocal(!isLocal);
+            isLocal = !isLocal;
+            populateToolbar();
+
+        });
+
+        if(edit) {
+            getToolbar().addCommandToOverflowMenu("Delete", null, evt -> {
+
                 r.delete();
                 new RuteList().showBack();
             });
-            cnt.add(b2);
-//        }
-//        cnt.add(multiToggleButton);
-        getToolbar().add(BorderLayout.EAST, cnt);
-
+        }
+        revalidate();
     }
 
 }
